phase_2_complete/langchain_fundamentals/chains_&_multistep/requirements.txt
# 🎯 Week 2: Days 14-16 - Sequential Chains & Memory
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-core>=0.1.0
openai>=1.3.0
tiktoken>=0.5.2
chromadb>=0.4.18
pypdf>=3.17.4
python-dotenv>=1.0.0
pydantic>=2.5.0
orjson>=3.9.10
rich>=13.7.0
colorama>=0.4.6
tqdm>=4.66.1
pytest>=7.4.3
black>=23.11.0
flake8>=6.1.0